Added interfaces for observer pattern (viewmodel->view notifications sending), added descriptions for javadoc. Connected history with calculation module (observer pattern). Added possibility of triggering the calculations. Moved DataTester to Model package (since business and validation logic shall happen in model, according to MS database).

+++ app/src/main/java/com/polsl/jwiiumlab6/viewmodel/ICalcResultListener.java
@@ -9,5 +9,5 @@ public interface ICalcResultListener {
      * Called by calculation performing object to retrieve calculations result.
      * @param resultValue Result value of the calculations.
      */
-    void retrieveResult(float resultValue);
+    void retrieveResult(double resultValue);
 }
